Create Codes
a CapsNet-based architecture called Speech-CapsNet (SCapsNet) is proposed for speech electroencephalograph (EEG) recording. In SCapsNet, the CapsNet architecture is modified using multi-level feature maps and multiple capsule layers. Moreover, the performance of the proposed method is compared with the baselines. SCapsNet's performance was assessed on two distinct datasets: the Imagined-Speech dataset and the Overt- and Imagined-Speech datasets.